Tratando latitude e longitude como string

+++ app/src/main/java/br/com/ufg/www/places/mvp/places/register/RegisterPlaceActivity.kt
@@ -28,8 +28,8 @@ class RegisterPlaceActivity : BaseActivity(), View.OnClickListener, Contract.Vie
             val description = editTextDescription.getString()
 
             val place = Place(
-                    latitude = latitude.toDouble(),
-                    longitude = longitude.toDouble(),
+                    latitude = latitude,
+                    longitude = longitude,
                     description = description,
                     login = App.userLoggedIn
             )
